Refactors and renames imageCount to getImageStatistics
